refactor: move config directory
