+++ backend/cmd/main.go
@@ -3,18 +3,18 @@ package main
 import (
 	"github.com/gin-gonic/gin"
 
-	"github.com/liviaruegger/MAC0350/backend/internal/config"
-	"github.com/liviaruegger/MAC0350/backend/internal/repository"
+	"github.com/liviaruegger/MAC0350/backend/config"
 	"github.com/liviaruegger/MAC0350/backend/internal/app"
 	"github.com/liviaruegger/MAC0350/backend/internal/handler"
+	"github.com/liviaruegger/MAC0350/backend/internal/repository"
 )
 
 func main() {
 	db := config.SetupDatabase()
 
-    userRepo := repository.NewUserRepository(db)
-    userService := app.NewUserService(userRepo)
-    userHandler := handler.NewUserHandler(userService)
+	userRepo := repository.NewUserRepository(db)
+	userService := app.NewUserService(userRepo)
+	userHandler := handler.NewUserHandler(userService)
 
 	router := gin.Default()
 	router.GET("/users/:id", userHandler.GetUserByID)
